+++ src/lib/utils/options/sources/typedoc.ts
@@ -540,6 +540,7 @@ export function addTypeDocOptions(options: Pick<Options, "addDeclaration">) {
             includeGroups: false,
             includeFolders: true,
             compactFolders: true,
+            includeReferences: true,
         },
     });